refactor: revise design pattern of intimacy calculator
* Facade pattern for `IntimacyCalculator` & `TextService`

* Abstract Factory pattern for `TextServiceFactory`

* Chain of Responsibility pattern for `ErrorHandler`

* Proxy pattern for `TextService`
