+++ fastapi/src/chatting/constants.py
@@ -1,1 +1,8 @@
+import os
+
+
 DEFAULT_INTIMACY: float = 36.5
+CLOVA_CLIENT_ID: str = os.environ.get('SNEK_CLOVA_CLIENT_ID')
+CLOVA_CLIENT_SECRET: str = os.environ.get('SNEK_CLOVA_CLIENT_SECRET')
+PAPAGO_CLIENT_ID: str = os.environ.get('SNEK_PAPAGO_CLIENT_ID')
+PAPAGO_CLIENT_SECRET: str = os.environ.get('SNEK_PAPAGO_CLIENT_SECRET')
