+++ fastapi/src/chatting/exceptions.py
@@ -1,4 +1,3 @@
-from typing import Any
 from fastapi import HTTPException
 
 from src.exceptions import ExternalApiError
@@ -29,6 +28,6 @@ class ChattingAlreadyExistException(HTTPException):
         super().__init__(400, detail='chatting already exist')
 
 
-class PapagoKoreanToKoreanException(HTTPException):
-    def __init__(self):
-        super().__init__(400, detail='papago does not support korean to korean translation')
+class KoreanTranslationException(ExternalApiError):
+    def __init__(self) -> None:
+        super().__init__(detail="does not support korean to korean translation")
