+++ fastapi/src/chatting/service.py
@@ -1,17 +1,11 @@
-from abc import ABCMeta, abstractmethod
-import os
-from typing import Dict, List, Tuple, Any
+from typing import List, Tuple
 
-import json
-import numpy as np
-import requests
 from sqlalchemy import insert, update, desc, or_, func
 from sqlalchemy.orm import Session as DbSession
 
-from src.chatting.constants import *
+from src.chatting.constants import DEFAULT_INTIMACY
 from src.chatting.exceptions import *
 from src.chatting.models import *
-from src.user.service import *
 
 
 def get_chatting_by_id(db: DbSession, chatting_id: int) -> Chatting:
@@ -36,12 +30,12 @@ def get_all_chattings(db: DbSession, user_id: int, is_approved: bool, limit: int
 
 
 def create_chatting(db: DbSession, user_id: int, responder_id: int) -> Chatting:
-    #if there is non-terminated chatting, throw exception
+    # if there is non-terminated chatting, throw exception
     chatting = db.query(Chatting).where(or_((Chatting.initiator_id == user_id) & (Chatting.responder_id == responder_id), (Chatting.responder_id == user_id) & (Chatting.initiator_id == responder_id))).where(
         Chatting.is_terminated == False).first()
     if chatting is not None:
         raise ChattingAlreadyExistException()
-    
+
     return db.scalar(
         insert(Chatting)
         .values(
@@ -169,6 +163,7 @@ def create_intimacy(db: DbSession, user_id: int | List[int], chatting_id: int, i
         "timestamp": timestamp,
     } for user_id in user_id]).returning(Intimacy)))
 
+
 def get_topics(db: DbSession, tag: str, limit: int, is_korean: bool = True) -> List[Topic]:
     topics = db.query(Topic).where(Topic.is_korean == is_korean).where(Topic.tag == tag).order_by(
         func.random()).limit(limit).all()
@@ -183,458 +178,3 @@ def intimacy_tag(intimacy: Intimacy | None) -> str:
         return "B"
     else:
         return "A"
-
-
-class ErrorHandler(metaclass=ABCMeta):
-    """An error handler for Clients."""
-    
-    @abstractmethod
-    def handle(self, response: requests.Response) -> None:
-        pass
-
-
-class Client(metaclass=ABCMeta):
-    """A client interface(abstract class) for external API calls."""
-
-    @abstractmethod
-    def api_url(self) -> str:
-        pass
-
-    @abstractmethod
-    def headers(self) -> Dict[str, str]:
-        pass
-
-    @abstractmethod
-    def data(self, text: str) -> str | Dict[str, str]:
-        pass
-
-    @abstractmethod
-    def error_handler(self) -> ErrorHandler:
-        pass
-
-    def post(self, text: str) -> requests.Response:
-        url = self.api_url()
-        headers = self.headers()
-        data = self.data(text)
-        return requests.post(url, data=data, headers=headers)
-
-    def request(self, text: str) -> Any:
-        response = self.post(text)
-        self.error_handler().handle(response)
-        return response.json()
-
-
-class TranslationClient(Client):
-    @abstractmethod
-    def parse_response(self, response: Any) -> str:
-        """Return translated_text"""
-
-    def translate(self, text: str) -> str:
-        response = self.request(text)
-        return self.parse_response(response)
-
-
-class SentimentClient(Client):
-    @abstractmethod
-    def parse_response(self, response: Any) -> float:
-        """Return sentiment"""
-
-    def get_sentiment(self, text: str) -> float:
-        response = self.request(text)
-        return self.parse_response(response)
-
-
-class PapagoClient(TranslationClient):
-    API_URL: str = "https://naveropenapi.apigw.ntruss.com/nmt/v1/translation"
-    CLIENT_ID: str = os.environ.get('SNEK_PAPAGO_CLIENT_ID')
-    CLIENT_SECRET: str = os.environ.get('SNEK_PAPAGO_CLIENT_SECRET')
-
-    @classmethod
-    def get_error_code(cls, response: requests.Response) -> String:
-        response = response.json()
-        return response['error']['errorCode']
-
-    @classmethod
-    def handle_error_response(cls, response: requests.Response) -> requests.Response:
-        if response.status_code != 200:
-            # errorCode == "N2MT05"
-            error_code = cls.get_error_code(response)
-            if error_code == "N2MT05":
-                raise PapagoKoreanToKoreanException()
-            else:
-                super().handle_error_response(response)
-        else:
-            return response
-
-    @classmethod
-    def translate(cls, text: str) -> str:
-        response = cls.post(text)
-        try:
-            response = cls.handle_error_response(response)
-        except (HTTPException, ExternalApiError):
-            return text
-        return cls.parse_response(response)
-
-    @classmethod
-    def api_error(cls) -> HTTPException:
-        return PapagoApiException()
-
-    @classmethod
-    def api_url(cls) -> str:
-        return cls.API_URL
-
-    @classmethod
-    def headers(cls) -> Dict[str, str]:
-        return {
-            "X-NCP-APIGW-API-KEY-ID": cls.CLIENT_ID,
-            "X-NCP-APIGW-API-KEY": cls.CLIENT_SECRET,
-        }
-
-    @classmethod
-    def data(cls, text: str) -> str | Dict[str, str]:
-        if len(text) > 999:
-            text = text[:999]
-        return {"source": "auto", "target": "ko", "text": text}
-
-    @classmethod
-    def parse_response(cls, response: requests.Response) -> str:
-        response = response.json()
-        return response["message"]["result"]["translatedText"]
-    
-
-def IgnoresEmptyInputTranslationClient(wrappee: Type[TranslationClient]) -> Type[TranslationClient]:
-    """
-    Proxy + singleton pattern that ignores empty input.
-    It creates an instance of level 2 type object (ABCMeta), which is level 1 type object.
-    """
-
-    class IgnoreEmptyInputClientImpl(WrappedClient(wrappee)):
-        @classmethod
-        def parse_response(cls, response: requests.Response) -> str:
-            return wrappee.parse_response(response)
-
-        @classmethod
-        def translate(cls, text: str) -> str:
-            if len(text) == 0:
-                return text
-            return wrappee.translate(text)
-
-    return IgnoreEmptyInputClientImpl
-
-
-class ClovaClient(SentimentClient):
-    API_URL: str = "https://naveropenapi.apigw.ntruss.com/sentiment-analysis/v1/analyze"
-    CLIENT_ID: str = os.environ.get('SNEK_CLOVA_CLIENT_ID')
-    CLIENT_SECRET: str = os.environ.get('SNEK_CLOVA_CLIENT_SECRET')
-
-    @classmethod
-    def api_error(cls) -> HTTPException:
-        return ClovaApiException()
-
-    @classmethod
-    def api_url(cls) -> str:
-        return cls.API_URL
-
-    @classmethod
-    def headers(cls) -> Dict[str, str]:
-        return {
-            "X-NCP-APIGW-API-KEY-ID": cls.CLIENT_ID,
-            "X-NCP-APIGW-API-KEY": cls.CLIENT_SECRET,
-            "Content-Type": "application/json",
-        }
-
-    @classmethod
-    def data(cls, text: str) -> str | Dict[str, str]:
-        return json.dumps({"content": text})
-
-    @classmethod
-    def parse_response(cls, response: requests.Response) -> float:
-        data = json.loads(response.text)
-        positive = data["document"]["confidence"]["positive"]
-        negative = data["document"]["confidence"]["negative"]
-        result = positive - negative
-        if result >= 0:
-            return result * 0.1
-        else:
-            return result * 0.05
-        # positive[0~100] negative[0~100]
-
-
-def IgnoresEmptyInputSentimentClient(wrappee: Type[SentimentClient]) -> Type[SentimentClient]:
-    """
-    Proxy + singleton pattern that ignores empty input.
-    It creates an instance of level 2 type object (ABCMeta), which is level 1 type object.
-    """
-
-    class IgnoreEmptyInputClientImpl(WrappedClient(wrappee)):
-        @classmethod
-        def parse_response(cls, response: requests.Response) -> float:
-            return wrappee.parse_response(response)
-        
-        @classmethod
-        def get_sentiment(cls, text: str) -> float:
-            if len(text) == 0:
-                return 0
-
-            return wrappee.get_sentiment(text)
-    
-    return IgnoreEmptyInputClientImpl
-
-
-class IntimacyCalculator:
-    __translation: Type[TranslationClient]
-    __sentiment: Type[SentimentClient]
-
-    def __init__(self, translation: Type[TranslationClient], sentiment: Type[SentimentClient]) -> None:
-        self.__translation = translation
-        self.__sentiment = sentiment
-
-    def calculate(
-        self,
-        user_id: int,
-        curr_texts: List[Text],
-        prev_texts: List[Text],
-        recent_intimacy: Intimacy,
-        similarity: float | None = None,
-        num_F: int | None = None,
-    ) -> float:
-        
-        curr_translated = self.__translation.translate('.'.join(text.msg for text in curr_texts))
-        sentiment = self.__sentiment.get_sentiment(curr_translated)
-
-        frequency = self.score_frequency(curr_texts)
-        frequency_delta = self.score_frequency_delta(prev_texts, curr_texts)
-        length = self.score_avg_length(curr_texts)
-        length_delta = self.score_avg_length_delta(prev_texts, curr_texts)
-        turn = self.score_turn(curr_texts, user_id)
-        turn_delta = self.score_turn_delta(prev_texts, curr_texts, user_id)
-        params = np.array([sentiment, frequency, frequency_delta,
-                          length, length_delta, turn, turn_delta])
-
-        weight = self.get_weight(similarity, num_F)
-        new_intimacy: float = recent_intimacy.intimacy + \
-            weight.dot(params.transpose())
-        return max(0, min(100, new_intimacy))
-
-    @staticmethod
-    def null_if_empty(func):
-        """Decorator pattern"""
-
-        def wrapper(texts: List[Text], *args, **kwargs) -> float | None:
-            if len(texts) == 0:
-                return None
-
-            return func(texts, *args, **kwargs)
-
-        return wrapper
-
-    @staticmethod
-    def get_delta(prev: float | None, curr: float | None) -> float | None:
-        if prev is None or curr is None:
-            return None
-        return curr - prev
-
-    @staticmethod
-    @null_if_empty
-    def get_frequency(texts: List[Text]) -> float | None:
-        """An average seconds for 20 texts"""
-
-        frequency = datetime.now() - texts[-1].timestamp
-        seconds = frequency.seconds + frequency.microseconds / 1e6
-        return seconds * 20 / len(texts)
-
-    @staticmethod
-    def get_frequency_delta(prev_text: List[Text], curr_text: List[Text]) -> float | None:
-        return IntimacyCalculator.get_delta(
-            IntimacyCalculator.get_frequency(prev_text),
-            IntimacyCalculator.get_frequency(curr_text))
-
-    @staticmethod
-    def score_frequency(texts: List[Text]) -> int:
-        seconds = IntimacyCalculator.get_frequency(texts)
-        if seconds is None:
-            return 0
-
-        if 3600 <= seconds:
-            return -5
-        elif 3000 <= seconds:
-            return -4
-        elif 2400 <= seconds:
-            return -2
-        elif 1800 <= seconds:
-            return 0
-        elif 1200 <= seconds:
-            return 3
-        elif 600 <= seconds:
-            return 5
-        else:
-            return 10
-
-    @staticmethod
-    def score_frequency_delta(prev_texts: List[Text], curr_texts: List[Text]) -> int:
-        seconds = IntimacyCalculator.get_frequency_delta(
-            prev_texts, curr_texts)
-        if seconds is None:
-            return 0
-
-        if 1800 <= seconds:
-            return -5
-        elif 1200 <= seconds:
-            return -3
-        elif 600 <= seconds:
-            return -1
-        elif -600 <= seconds:
-            return 0
-        elif -1200 <= seconds:
-            return 3
-        elif -1800 <= seconds:
-            return 5
-        else:
-            return 10
-
-    @staticmethod
-    @null_if_empty
-    def get_avg_length(texts: List[Text]) -> float | None:
-        avg_len = 0
-        for text in texts:
-            avg_len += len(text.msg)
-        avg_len /= len(texts)
-        return avg_len
-
-    @staticmethod
-    def get_avg_length_delta(
-        prev_texts: List[Text], curr_texts: List[Text]
-    ) -> float | None:
-        return IntimacyCalculator.get_delta(
-            IntimacyCalculator.get_avg_length(prev_texts),
-            IntimacyCalculator.get_avg_length(curr_texts))
-
-    @staticmethod
-    def score_avg_length(texts: List[Text]) -> int:
-        avg_len = IntimacyCalculator.get_avg_length(texts)
-        if avg_len is None:
-            return 0
-
-        if avg_len >= 30:
-            return 10
-        elif avg_len >= 20:
-            return 7
-        elif avg_len >= 10:
-            return 4
-        elif avg_len >= 5:
-            return 0
-        else:
-            return -5
-
-    @staticmethod
-    def score_avg_length_delta(prev_texts: List[Text], curr_texts: List[Text]) -> int:
-        avg_len = IntimacyCalculator.get_avg_length_delta(
-            prev_texts, curr_texts)
-        if avg_len is None:
-            return 0
-
-        if avg_len >= 15:
-            return 10
-        elif avg_len >= 10:
-            return 5
-        elif avg_len >= -10:
-            return 0
-        elif avg_len >= -20:
-            return -3
-        else:
-            return -5
-
-    # scope of rate[0,1]
-    @staticmethod
-    @null_if_empty
-    def get_turn(texts: List[Text], user_id: int) -> float | None:
-        """How many turns that the user took among provided texts"""
-
-        turn = 0
-        for text in texts:
-            if text.sender_id == user_id:
-                turn += 1
-
-        rate = turn / len(texts)
-        return rate
-
-    # 작을수록 개선된 것(0.5에 가까워졌으므로)
-    @staticmethod
-    def get_turn_delta(
-        prev_texts: List[Text], curr_texts: List[Text], user_id: int
-    ) -> float | None:
-        prev_rate = IntimacyCalculator.get_turn(prev_texts, user_id)
-        curr_rate = IntimacyCalculator.get_turn(curr_texts, user_id)
-
-        if prev_rate is None or curr_rate is None:
-            return None
-
-        return abs(curr_rate - 0.5) - abs(prev_rate - 0.5)
-
-    @staticmethod
-    def score_turn(texts: List[Text], user_id: int) -> int:
-        rate = IntimacyCalculator.get_turn(texts, user_id)
-        if rate is None:
-            return 0
-
-        if 0.4 <= rate <= 0.6:
-            return 10
-        elif 0.3 <= rate <= 0.7:
-            return 5
-        elif 0.2 <= rate <= 0.8:
-            return 0
-        elif 0.1 <= rate <= 0.9:
-            return -3
-        else:
-            return -5
-
-    @staticmethod
-    def score_turn_delta(
-        prev_texts: List[Text], curr_texts: List[Text], user_id: int
-    ) -> int:
-        rate = IntimacyCalculator.get_turn_delta(
-            prev_texts, curr_texts, user_id)
-        if rate is None:
-            return 0
-
-        if 0.3 <= rate:
-            return -5
-        elif 0.1 <= rate:
-            return -3
-        elif 0.0 <= rate:
-            return 0
-        else:
-            return 10
-
-    @staticmethod
-    def get_weight(similarity: float | None = None, num_F: int | None = None) -> np.ndarray[float]:
-        # Default
-        if similarity is None or num_F is None:
-            return np.array([0.1, 0.3, 0, 0.3, 0, 0.3, 0])
-
-        if similarity < 0.2:
-            # sentiment, frequency, frequency_delta, length, length_delta, turn, turn_delta
-            weight = np.array([0.1, 0.2, 0.1, 0.2, 0.1, 0.2, 0.1])
-        elif similarity < 0.4:
-            weight = np.array([0.1, 0.19, 0.11, 0.19, 0.11, 0.19, 0.11])
-        elif similarity < 0.6:
-            weight = np.array([0.1, 0.18, 0.12, 0.18, 0.12, 0.18, 0.12])
-        elif similarity < 0.8:
-            weight = np.array([0.1, 0.17, 0.13, 0.17, 0.13, 0.17, 0.13])
-        else:
-            weight = np.array([0.1, 0.16, 0.14, 0.16, 0.14, 0.16, 0.14])
-
-        if num_F == 0:
-            weight[0] += 0.03
-            for i in range(1, 7):
-                weight[i] -= 0.005
-        elif num_F == 1:
-            weight[0] += 0.06
-            for i in range(1, 7):
-                weight[i] -= 0.01
-        elif num_F == 2:
-            weight[0] += 0.09
-            for i in range(1, 7):
-                weight[i] -= 0.015
-
-        return weight
